refactor(web): group route boundary helpers into components/route/
Move route-error/route-loading into components/route/ (error.tsx, loading.tsx); update 6 imports.

Matches the repo's by-domain convention (sidebar/, blog/, navbar/); git mv preserves history.

Pure move, no behavior change. Rationale in .github/audit/2026-06-29-component-organization.md.
