+++ web/next/src/app/(console)/console/loading.tsx
@@ -1,4 +1,4 @@
-import { RouteLoading } from "@/components/route-loading"
+import { RouteLoading } from "@/components/route/loading"
 
 export default function Loading() {
   return <RouteLoading className="flex-1" />
