+++ web/next/src/app/error.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { RouteError } from "@/components/route-error"
+import { RouteError } from "@/components/route/error"
 
 export default function Error(props: { error: Error & { digest?: string }; reset: () => void }) {
   return <RouteError {...props} className="min-h-svh" />
